+++ frontend/src/lib/components/utils/LoadingSpinner.svelte
